Add files via upload
These 4 matlab files describe the Monte Carlo estimation of uncertainty. The two "m-files" are MonteCarloNMRdilution.m which has the main program and nmraggfitbind.m which is an associated file for the data fitting process. The two "mat" datafiles are nmrdilutiondata.mat and MonteCarloResults.mat (large). To use in matlab: i) load the input data in the workspace with ->load nmrdilutiondata, then execute the Monte Carlo by -> MonteCarloNMRdilution After a few minutes (change cycln=20000 in line 181 to make it longer or shorter) the simulation will be finished and the output in the workspace should look similar to the "MonteCarloResults" shown here (you do have to save your own output).